fix(server): fix the repositories

+++ server/src/db/sequalize/category/category.repository.ts
@@ -8,7 +8,7 @@ export class CategoryRepository implements Repository<Category> {
     //#region Create
 
     /**
-     * Add a category to the repository.
+     * Create a category to the repository
      * @param category category
      * @returns Promise<Category>.
      */
@@ -19,43 +19,79 @@ export class CategoryRepository implements Repository<Category> {
             );
         }
 
-        const model = await ModelCategory.create(category);
-
-        return model.get({ plain: true });
+        return (await ModelCategory.create(category))?.get();
     }
 
     //#endregion
 
     //#region Read
 
-    readAll(): Promise<Category[]> {
-        return ModelCategory.findAll();
+    /**
+     * Reads all categories from the repository
+     * @returns Promise<Category[]>
+     */
+    async readAll(): Promise<Category[]> {
+        return (await ModelCategory.findAll()).map(model => model.get());
     }
 
-    readById(id: number): Promise<Category | null> {
-        return ModelCategory.findOne({ where: { id } });
+    /**
+     * Reads a category from the repository with the id
+     * @param id id 
+     * @returns 
+     */
+    async readById(id: number): Promise<Category | undefined> {
+        if (!id || typeof id !== 'number' || isNaN(id)) {
+            return undefined;
+        }
+
+        return (await ModelCategory.findOne({ where: { id } }))?.get();
     }
 
     //#endregion
 
     //#region Update
 
+    /**
+     * Updates a category from the repository with the id
+     * @param id 
+     * @param values 
+     * @returns 
+     */
     async update(
         id: number,
         values: Partial<Omit<Category, 'id'>> = {}
-    ): Promise<number | undefined> {
-        return (await ModelCategory.update(values, { where: { id } })).at(0);
+    ): Promise<number> {
+        if (!id || typeof id !== 'number' || isNaN(id)) {
+            return 0;
+        }
+
+        const [countAffected] = await ModelCategory.update(values, { where: { id } });
+        
+        return countAffected;
     }
 
     //#endregion
 
     //#region Delete
 
+    /**
+     * Removes all categories from the repository
+     * @returns 
+     */
     deleteAll(): Promise<number> {
         return ModelCategory.destroy();
     }
 
+    /**
+     * Removes a catogory from the repository with the id
+     * @param id 
+     * @returns 
+     */
     deleteById(id: number): Promise<number> {
+        if (!id || typeof id !== 'number' || isNaN(id)) {
+            return Promise.resolve(0);
+        }
+
         return ModelCategory.destroy({ where: { id } });
     }
